+++ R/qol.R
@@ -48,7 +48,7 @@
 #'
 #' Generate dummy data: [dummy_data()]
 #'
-#' Small helpers: [inverse()], [setcolorder_by_pattern()],
+#' Small helpers: [libname()], [inverse()], [setcolorder_by_pattern()],
 #' [drop_type_vars()], [fuse_variables()], [get_excel_range()], [replace_except()].
 #'
 #' Split data frame: [split_by_var()], [split_by_condition()].